requirements.txt
telethon
python-dotenv
tqdm
